refactor: split worker tests
