+++ test/worker-init.test.ts
@@ -1,0 +1,734 @@
+import { vi, describe, it, expect, beforeEach, afterEach } from "vitest";
+import {
+  EncoderConfig,
+  InitializeWorkerMessage,
+  FinalizeWorkerMessage,
+  CancelWorkerMessage,
+  AddAudioDataMessage,
+  MainThreadMessage,
+} from "../src/types";
+import {
+  setupGlobals,
+  cleanupGlobals,
+  resetMocks,
+  importWorker,
+  mockSelf,
+  mockMuxerInstanceForWorker,
+} from "./helpers/worker-test-utils";
+import { Mp4MuxerWrapper as ActualMp4MuxerWrapper } from "../src/mp4muxer";
+
+describe("worker", () => {
+  let config: EncoderConfig;
+  let Mp4MuxerWrapperMock: ReturnType<
+    typeof vi.mocked<typeof ActualMp4MuxerWrapper>
+  >;
+
+  beforeEach(async () => {
+    vi.resetModules();
+    setupGlobals();
+    resetMocks();
+    const mp4muxerModule = await import("../src/mp4muxer");
+    Mp4MuxerWrapperMock = vi.mocked(mp4muxerModule.Mp4MuxerWrapper);
+    Mp4MuxerWrapperMock.mockImplementation(() => mockMuxerInstanceForWorker as any);
+    await importWorker();
+    config = {
+      width: 640,
+      height: 480,
+      frameRate: 30,
+      videoBitrate: 1000000,
+      audioBitrate: 128000,
+      sampleRate: 48000,
+      channels: 2,
+      codec: { video: "avc", audio: "aac" },
+      container: "mp4",
+      latencyMode: "quality",
+    };
+  });
+
+  afterEach(() => {
+    cleanupGlobals();
+  });
+
+  it("initializes and finalizes", async () => {
+    if (!global.self.onmessage) {
+      throw new Error("Worker onmessage handler not set up by script import");
+    }
+
+    const expectedOutputArray = new Uint8Array([1, 2, 3, 4]);
+    // Mock muxer.finalize to return the Uint8Array directly
+    mockMuxerInstanceForWorker.finalize = vi.fn(() => expectedOutputArray);
+
+    // Simulate receiving an initialize message
+    const initMessage: InitializeWorkerMessage = { type: "initialize", config };
+    mockSelf.postMessage.mockClear();
+    await global.self.onmessage({ data: initMessage } as MessageEvent);
+    expect(mockSelf.postMessage).toHaveBeenCalledWith(
+      {
+        type: "initialized",
+        actualVideoCodec: "avc1.42001f",
+        actualAudioCodec: "mp4a.40.2",
+      },
+      undefined,
+    );
+
+    // Simulate receiving a finalize message
+    mockSelf.postMessage.mockClear();
+    const finalizeMessage: FinalizeWorkerMessage = { type: "finalize" };
+    await global.self.onmessage({ data: finalizeMessage } as MessageEvent);
+
+    expect(mockSelf.postMessage).toHaveBeenCalledTimes(1);
+    const finalizedCall = mockSelf.postMessage.mock.calls[0];
+    const msg = finalizedCall[0];
+
+    expect(msg.type).toBe("finalized");
+    expect(msg.output).toBeInstanceOf(Uint8Array);
+    expect(msg.output).toEqual(expectedOutputArray); // Compare Uint8Array content
+
+    expect(finalizedCall[1]).toBeInstanceOf(Array);
+    expect(finalizedCall[1].length).toBe(1);
+    const transferredObject = finalizedCall[1][0];
+    expect(transferredObject).toBeInstanceOf(ArrayBuffer);
+    // Ensure the transferred ArrayBuffer has the same content as the output Uint8Array's buffer
+    expect(new Uint8Array(transferredObject as ArrayBuffer)).toEqual(
+      expectedOutputArray,
+    );
+    // The transferred ArrayBuffer should be the .buffer of the msg.output Uint8Array
+    expect(transferredObject).toBe(expectedOutputArray.buffer);
+  });
+
+  it("connects audio port and handles messages", async () => {
+    if (!global.self.onmessage)
+      throw new Error("Worker onmessage handler not set up");
+    const port: any = { postMessage: vi.fn(), onmessage: null, close: vi.fn() };
+    const initMessage: InitializeWorkerMessage = { type: "initialize", config };
+    await global.self.onmessage({ data: initMessage } as MessageEvent);
+    const connectMsg = { type: "connectAudioPort", port } as any;
+    await global.self.onmessage({ data: connectMsg } as MessageEvent);
+    expect(port.onmessage).toBeTypeOf("function");
+    mockSelf.AudioEncoder.mock.results[0].value.encode.mockClear();
+
+    const addAudioMsg: AddAudioDataMessage = {
+      type: "addAudioData",
+      audioData: [new Float32Array(1), new Float32Array(1)],
+      timestamp: 0,
+      format: "f32-planar",
+      sampleRate: config.sampleRate,
+      numberOfFrames: 1,
+      numberOfChannels: 2,
+    };
+    if (port.onmessage)
+      await port.onmessage({ data: addAudioMsg } as MessageEvent);
+    expect(
+      mockSelf.AudioEncoder.mock.results[0].value.encode,
+    ).toHaveBeenCalled();
+  });
+
+  it("handles cancel message", async () => {
+    if (!global.self.onmessage) {
+      throw new Error("Worker onmessage handler not set up by script import");
+    }
+    // Initialize first (or part of it, enough to set up for cancel)
+    const initMessage: InitializeWorkerMessage = { type: "initialize", config };
+    await global.self.onmessage({ data: initMessage } as MessageEvent);
+    expect(mockSelf.postMessage).toHaveBeenCalledWith(
+      {
+        type: "initialized",
+        actualVideoCodec: "avc1.42001f",
+        actualAudioCodec: "mp4a.40.2",
+      },
+      undefined,
+    );
+    mockSelf.postMessage.mockClear();
+
+    // Simulate receiving a cancel message
+    const cancelMessage: CancelWorkerMessage = { type: "cancel" };
+    await global.self.onmessage({ data: cancelMessage } as MessageEvent);
+    expect(mockSelf.postMessage).toHaveBeenCalledWith(
+      { type: "cancelled" },
+      undefined,
+    );
+
+    // After cancellation, other messages should be ignored
+    mockSelf.postMessage.mockClear();
+    const finalizeMessage: FinalizeWorkerMessage = { type: "finalize" };
+    await global.self.onmessage({ data: finalizeMessage } as MessageEvent);
+    expect(mockSelf.postMessage).not.toHaveBeenCalled();
+
+    // Reinitialize should reset the cancelled state
+    const initMessage2: InitializeWorkerMessage = {
+      type: "initialize",
+      config,
+    };
+    await global.self.onmessage({ data: initMessage2 } as MessageEvent);
+    expect(mockSelf.postMessage).toHaveBeenCalledWith(
+      {
+        type: "initialized",
+        actualVideoCodec: "avc1.42001f",
+        actualAudioCodec: "mp4a.40.2",
+      },
+      undefined,
+    );
+  });
+
+  // Add more tests for addVideoData, addAudioData, error handling, etc.
+  describe("worker error handling during initialization", () => {
+    it("should post an error if video codec is not supported", async () => {
+      if (!global.self.onmessage)
+        throw new Error("Worker onmessage handler not set up");
+
+      // @ts-ignore
+      mockSelf.VideoEncoder.isConfigSupported = vi.fn(() =>
+        Promise.resolve({ supported: false, config: null }),
+      );
+
+      const initMessage: InitializeWorkerMessage = {
+        type: "initialize",
+        config,
+      };
+      await global.self.onmessage({ data: initMessage } as MessageEvent);
+
+      expect(mockSelf.postMessage).toHaveBeenCalledWith(
+        {
+          type: "error",
+          errorDetail: {
+            message: "Worker: Video codec avc config not supported.",
+            type: "not-supported",
+          },
+        },
+        undefined,
+      );
+    });
+
+    it("should post an error if audio codec is not supported", async () => {
+      if (!global.self.onmessage)
+        throw new Error("Worker onmessage handler not set up");
+
+      // @ts-ignore
+      mockSelf.AudioEncoder.isConfigSupported = vi.fn(() =>
+        Promise.resolve({ supported: false, config: null }),
+      );
+
+      const initMessage: InitializeWorkerMessage = {
+        type: "initialize",
+        config,
+      };
+      await global.self.onmessage({ data: initMessage } as MessageEvent);
+
+      expect(mockSelf.postMessage).toHaveBeenCalledWith(
+        {
+          type: "error",
+          errorDetail: {
+            message: "Worker: Audio codec aac config not supported.",
+            type: "not-supported",
+          },
+        },
+        undefined,
+      );
+    });
+
+    it("should post an error if video encoder configuration fails", async () => {
+      if (!global.self.onmessage)
+        throw new Error("Worker onmessage handler not set up");
+
+      const configureError = new Error("Video configure failed");
+      // @ts-ignore
+      mockSelf.VideoEncoder = vi.fn(() => ({
+        configure: vi.fn().mockImplementation(() => {
+          throw configureError;
+        }),
+        close: vi.fn(),
+        state: "unconfigured",
+      }));
+      // @ts-ignore
+      mockSelf.VideoEncoder.isConfigSupported = vi.fn(() =>
+        Promise.resolve({ supported: true, config: { codec: "avc1.42001f" } }),
+      );
+      globalThis.VideoEncoder = mockSelf.VideoEncoder;
+
+      const initMessage: InitializeWorkerMessage = {
+        type: "initialize",
+        config,
+      };
+      await global.self.onmessage({ data: initMessage } as MessageEvent);
+
+      expect(mockSelf.postMessage).toHaveBeenCalledWith(
+        {
+          type: "error",
+          errorDetail: {
+            message:
+              "Worker: Failed to initialize VideoEncoder: Video configure failed",
+            type: "initialization-failed",
+            stack: expect.any(String),
+          },
+        },
+        undefined,
+      );
+    });
+
+    it("should post an error if audio encoder configuration fails", async () => {
+      if (!global.self.onmessage)
+        throw new Error("Worker onmessage handler not set up");
+
+      const configureError = new Error("Audio configure failed");
+      // @ts-ignore
+      mockSelf.AudioEncoder = vi.fn(() => ({
+        configure: vi.fn().mockImplementation(() => {
+          throw configureError;
+        }),
+        close: vi.fn(),
+        state: "unconfigured",
+      }));
+      // @ts-ignore
+      mockSelf.AudioEncoder.isConfigSupported = vi.fn(() =>
+        Promise.resolve({ supported: true, config: { codec: "mp4a.40.2" } }),
+      );
+      globalThis.AudioEncoder = mockSelf.AudioEncoder;
+
+      const initMessage: InitializeWorkerMessage = {
+        type: "initialize",
+        config,
+      };
+      await global.self.onmessage({ data: initMessage } as MessageEvent);
+
+      expect(mockSelf.postMessage).toHaveBeenCalledWith(
+        {
+          type: "error",
+          errorDetail: {
+            message:
+              "Worker: Failed to initialize AudioEncoder: Audio configure failed",
+            type: "initialization-failed",
+            stack: expect.any(String),
+          },
+        },
+        undefined,
+      );
+    });
+
+    it("should post an error if config is missing", async () => {
+      if (!global.self.onmessage)
+        throw new Error("Worker onmessage handler not set up");
+      // @ts-ignore
+      const initMessage: InitializeWorkerMessage = {
+        type: "initialize",
+        config: null as any,
+      };
+      await global.self.onmessage({ data: initMessage } as MessageEvent);
+      expect(mockSelf.postMessage).toHaveBeenCalledWith(
+        {
+          type: "error",
+          errorDetail: {
+            message: "Worker: Configuration is missing.",
+            type: "initialization-failed",
+          },
+        },
+        undefined,
+      );
+    });
+
+    it("should post an error if webm container is specified", async () => {
+      if (!global.self.onmessage)
+        throw new Error("Worker onmessage handler not set up");
+      const webmConfig = { ...config, container: "webm" as const };
+      const initMessage: InitializeWorkerMessage = {
+        type: "initialize",
+        config: webmConfig,
+      };
+      await global.self.onmessage({ data: initMessage } as MessageEvent);
+      expect(mockSelf.postMessage).toHaveBeenCalledWith(
+        {
+          type: "error",
+          errorDetail: {
+            message: "Worker: WebM container is not supported in this version.",
+            type: "not-supported",
+          },
+        },
+        undefined,
+      );
+    });
+
+    it("should fallback to avc if preferred video codec (vp9) is not supported but avc is", async () => {
+      if (!global.self.onmessage)
+        throw new Error("Worker onmessage handler not set up");
+      const consoleWarnSpy = vi
+        .spyOn(console, "warn")
+        .mockImplementation(() => {});
+      const vp9Config = {
+        ...config,
+        codec: { ...config.codec, video: "vp9" as const },
+      };
+
+      // @ts-ignore
+      mockSelf.VideoEncoder.isConfigSupported = vi.fn(async (_cfg) => {
+        if (_cfg.codec.startsWith("vp09"))
+          return { supported: false, config: null };
+        if (_cfg.codec.startsWith("avc1"))
+          return {
+            supported: true,
+            config: { ..._cfg, codec: "avc1.42001f.test" },
+          };
+        return { supported: false, config: null };
+      });
+      globalThis.VideoEncoder = mockSelf.VideoEncoder;
+
+      const initMessage: InitializeWorkerMessage = {
+        type: "initialize",
+        config: vp9Config,
+      };
+      await global.self.onmessage({ data: initMessage } as MessageEvent);
+
+      expect(consoleWarnSpy).toHaveBeenCalledWith(
+        "Worker: Video codec vp9 not supported or config invalid. Falling back to AVC.",
+      );
+      expect(mockSelf.postMessage).toHaveBeenCalledWith(
+        expect.objectContaining({
+          type: "initialized",
+          actualVideoCodec: "avc1.42001f.test",
+        }),
+        undefined,
+      );
+      consoleWarnSpy.mockRestore();
+    });
+
+    it("should post error if fallback video codec (avc) is also not supported", async () => {
+      if (!global.self.onmessage)
+        throw new Error("Worker onmessage handler not set up");
+      const vp9Config = {
+        ...config,
+        codec: { ...config.codec, video: "vp9" as const },
+      };
+
+      // @ts-ignore
+      mockSelf.VideoEncoder.isConfigSupported = vi.fn(async (_cfg) => {
+        return { supported: false, config: null }; // All codecs unsupported
+      });
+      globalThis.VideoEncoder = mockSelf.VideoEncoder;
+
+      const initMessage: InitializeWorkerMessage = {
+        type: "initialize",
+        config: vp9Config,
+      };
+      await global.self.onmessage({ data: initMessage } as MessageEvent);
+
+      expect(mockSelf.postMessage).toHaveBeenCalledWith(
+        {
+          type: "error",
+          errorDetail: {
+            message:
+              "Worker: AVC (H.264) video codec is not supported after fallback.",
+            type: "not-supported",
+          },
+        },
+        undefined,
+      );
+    });
+
+    it("should fallback to aac if preferred audio codec (opus) is not supported but aac is", async () => {
+      if (!global.self.onmessage)
+        throw new Error("Worker onmessage handler not set up");
+      const consoleWarnSpy = vi
+        .spyOn(console, "warn")
+        .mockImplementation(() => {});
+      const opusConfig = {
+        ...config,
+        codec: { ...config.codec, audio: "opus" as const },
+      };
+
+      // @ts-ignore
+      mockSelf.AudioEncoder.isConfigSupported = vi.fn(async (_cfg) => {
+        if (_cfg.codec === "opus") return { supported: false, config: null };
+        if (_cfg.codec === "mp4a.40.2")
+          return {
+            supported: true,
+            config: { ..._cfg, codec: "mp4a.40.2.test" },
+          };
+        return { supported: false, config: null };
+      });
+      globalThis.AudioEncoder = mockSelf.AudioEncoder;
+
+      const initMessage: InitializeWorkerMessage = {
+        type: "initialize",
+        config: opusConfig,
+      };
+      await global.self.onmessage({ data: initMessage } as MessageEvent);
+      expect(consoleWarnSpy).toHaveBeenCalledWith(
+        "Worker: Audio codec opus not supported or config invalid. Falling back to AAC.",
+      );
+      expect(mockSelf.postMessage).toHaveBeenCalledWith(
+        expect.objectContaining({
+          type: "initialized",
+          actualAudioCodec: "mp4a.40.2.test",
+        }),
+        undefined,
+      );
+      consoleWarnSpy.mockRestore();
+    });
+
+    it("should post error if fallback audio codec (aac) is also not supported", async () => {
+      if (!global.self.onmessage)
+        throw new Error("Worker onmessage handler not set up");
+      const opusConfig = {
+        ...config,
+        codec: { ...config.codec, audio: "opus" as const },
+      };
+      // @ts-ignore
+      mockSelf.AudioEncoder.isConfigSupported = vi.fn(async (_cfg) => {
+        return { supported: false, config: null }; // All codecs unsupported
+      });
+      globalThis.AudioEncoder = mockSelf.AudioEncoder;
+
+      const initMessage: InitializeWorkerMessage = {
+        type: "initialize",
+        config: opusConfig,
+      };
+      await global.self.onmessage({ data: initMessage } as MessageEvent);
+
+      expect(mockSelf.postMessage).toHaveBeenCalledWith(
+        {
+          type: "error",
+          errorDetail: {
+            message: "Worker: AAC audio codec is not supported after fallback.",
+            type: "not-supported",
+          },
+        },
+        undefined,
+      );
+    });
+
+    it("should post error if VideoEncoder API is not available", async () => {
+      if (!global.self.onmessage)
+        throw new Error("Worker onmessage handler not set up");
+      const originalVideoEncoder = globalThis.VideoEncoder;
+      delete (globalThis as any).VideoEncoder;
+      delete (mockSelf as any).VideoEncoder; // Also remove from mockSelf
+
+      const initMessage: InitializeWorkerMessage = {
+        type: "initialize",
+        config,
+      };
+      await global.self.onmessage({ data: initMessage } as MessageEvent);
+      expect(mockSelf.postMessage).toHaveBeenCalledWith(
+        {
+          type: "error",
+          errorDetail: {
+            message: "Worker: VideoEncoder not available",
+            type: "not-supported",
+          },
+        },
+        undefined,
+      );
+      globalThis.VideoEncoder = originalVideoEncoder; // Restore
+      // @ts-ignore
+      mockSelf.VideoEncoder = vi.fn(() => ({
+        configure: vi.fn(),
+        encode: vi.fn(),
+        flush: vi.fn(),
+        close: vi.fn(),
+        state: "unconfigured",
+      }));
+      // @ts-ignore
+      mockSelf.VideoEncoder.isConfigSupported = vi.fn(() =>
+        Promise.resolve({ supported: true, config: { codec: "avc1.42001f" } }),
+      );
+    });
+
+    it("should post error if AudioEncoder API is not available", async () => {
+      if (!global.self.onmessage)
+        throw new Error("Worker onmessage handler not set up");
+      const originalAudioEncoder = globalThis.AudioEncoder;
+      delete (globalThis as any).AudioEncoder;
+      delete (mockSelf as any).AudioEncoder;
+
+      const initMessage: InitializeWorkerMessage = {
+        type: "initialize",
+        config,
+      };
+      await global.self.onmessage({ data: initMessage } as MessageEvent);
+      expect(mockSelf.postMessage).toHaveBeenCalledWith(
+        {
+          type: "error",
+          errorDetail: {
+            message: "Worker: AudioEncoder not available",
+            type: "not-supported",
+          },
+        },
+        undefined,
+      );
+      globalThis.AudioEncoder = originalAudioEncoder; // Restore
+      // @ts-ignore
+      mockSelf.AudioEncoder = vi.fn(() => ({
+        configure: vi.fn(),
+        encode: vi.fn(),
+        flush: vi.fn(),
+        close: vi.fn(),
+        state: "unconfigured",
+      }));
+      // @ts-ignore
+      mockSelf.AudioEncoder.isConfigSupported = vi.fn(() =>
+        Promise.resolve({ supported: true, config: { codec: "mp4a.40.2" } }),
+      );
+    });
+
+    it("should post error if VideoEncoder constructor throws", async () => {
+      if (!global.self.onmessage)
+        throw new Error("Worker onmessage handler not set up");
+      const originalVideoEncoderCtor = mockSelf.VideoEncoder;
+      const constructorError = new Error("Video Constructor failed");
+
+      const mockVideoEncoderConstructor = vi.fn();
+      // @ts-ignore
+      mockVideoEncoderConstructor.isConfigSupported = vi.fn(() =>
+        Promise.resolve({
+          supported: true,
+          config: { codec: "avc1.42001f.test" },
+        }),
+      );
+      mockVideoEncoderConstructor.mockImplementation(() => {
+        throw constructorError;
+      });
+
+      mockSelf.VideoEncoder = mockVideoEncoderConstructor;
+      globalThis.VideoEncoder = mockSelf.VideoEncoder;
+
+      const originalAudioEncoderCtor = mockSelf.AudioEncoder;
+      // @ts-ignore
+      const MockAudioEncoderInstance = {
+        configure: vi.fn(),
+        encode: vi.fn(),
+        flush: vi.fn(),
+        close: vi.fn(),
+        state: "unconfigured",
+      };
+      // @ts-ignore
+      mockSelf.AudioEncoder = vi.fn(() => MockAudioEncoderInstance);
+      // @ts-ignore
+      mockSelf.AudioEncoder.isConfigSupported = vi.fn(() =>
+        Promise.resolve({ supported: true, config: { codec: "mp4a.40.2" } }),
+      );
+      globalThis.AudioEncoder = mockSelf.AudioEncoder;
+
+      const initMessage: InitializeWorkerMessage = {
+        type: "initialize",
+        config,
+      };
+      await global.self.onmessage({ data: initMessage } as MessageEvent);
+
+      expect(mockSelf.postMessage).toHaveBeenCalledWith(
+        expect.objectContaining({
+          type: "error",
+          errorDetail: expect.objectContaining({
+            message: `Worker: Failed to initialize VideoEncoder: ${constructorError.message}`,
+            type: "initialization-failed",
+            stack: expect.any(String),
+          }),
+        }),
+        undefined,
+      );
+
+      mockSelf.VideoEncoder = originalVideoEncoderCtor;
+      globalThis.VideoEncoder = mockSelf.VideoEncoder;
+      mockSelf.AudioEncoder = originalAudioEncoderCtor;
+      globalThis.AudioEncoder = mockSelf.AudioEncoder;
+    });
+
+    it("should post error if AudioEncoder constructor throws", async () => {
+      if (!global.self.onmessage)
+        throw new Error("Worker onmessage handler not set up");
+
+      // === リセットフェーズ ===
+      const originalVideoEncoderCtorForThisTest = mockSelf.VideoEncoder;
+      const originalAudioEncoderCtorForThisTest = mockSelf.AudioEncoder;
+
+      const MockVideoInstance = {
+        configure: vi.fn(),
+        encode: vi.fn(),
+        flush: vi.fn(),
+        close: vi.fn(),
+        state: "unconfigured",
+      };
+      const MockAudioInstance = {
+        configure: vi.fn(),
+        encode: vi.fn(),
+        flush: vi.fn(),
+        close: vi.fn(),
+        state: "unconfigured",
+      };
+
+      // @ts-ignore
+      mockSelf.VideoEncoder = vi.fn(() => MockVideoInstance);
+      // @ts-ignore
+      mockSelf.VideoEncoder.isConfigSupported = vi.fn(() =>
+        Promise.resolve({
+          supported: true,
+          config: { codec: "avc1.42001f.default" },
+        }),
+      );
+      globalThis.VideoEncoder = mockSelf.VideoEncoder;
+
+      // @ts-ignore
+      mockSelf.AudioEncoder = vi.fn(() => MockAudioInstance);
+      // @ts-ignore
+      mockSelf.AudioEncoder.isConfigSupported = vi.fn(() =>
+        Promise.resolve({
+          supported: true,
+          config: { codec: "mp4a.40.2.default" },
+        }),
+      );
+      globalThis.AudioEncoder = mockSelf.AudioEncoder;
+      // === リセットフェーズここまで ===
+
+      // === VideoEncoder をこのテスト用に正常動作させる設定 ===
+      // @ts-ignore
+      mockSelf.VideoEncoder.isConfigSupported = vi.fn(() =>
+        Promise.resolve({
+          supported: true,
+          config: { codec: "avc1.42001f.test" },
+        }),
+      );
+      // @ts-ignore
+      mockSelf.VideoEncoder.mockImplementation(() => MockVideoInstance);
+      globalThis.VideoEncoder = mockSelf.VideoEncoder;
+
+      // === AudioEncoder がエラーを投げるように設定 ===
+      const constructorError = new Error("Audio Constructor failed for test");
+      const mockAudioEncoderConstructorThatThrows = vi.fn();
+      // @ts-ignore
+      mockAudioEncoderConstructorThatThrows.isConfigSupported = vi.fn(() =>
+        Promise.resolve({
+          supported: true,
+          config: { codec: "mp4a.40.2.test" },
+        }),
+      );
+      mockAudioEncoderConstructorThatThrows.mockImplementation(() => {
+        throw constructorError;
+      });
+
+      mockSelf.AudioEncoder = mockAudioEncoderConstructorThatThrows;
+      globalThis.AudioEncoder = mockSelf.AudioEncoder;
+
+      const initMessage: InitializeWorkerMessage = {
+        type: "initialize",
+        config,
+      };
+      await global.self.onmessage({ data: initMessage } as MessageEvent);
+
+      expect(mockSelf.postMessage).toHaveBeenCalledWith(
+        expect.objectContaining({
+          type: "error",
+          errorDetail: expect.objectContaining({
+            message: `Worker: Failed to initialize AudioEncoder: ${constructorError.message}`,
+            type: "initialization-failed",
+            stack: expect.any(String),
+          }),
+        }),
+        undefined,
+      );
+
+      // リストア
+      mockSelf.VideoEncoder = originalVideoEncoderCtorForThisTest;
+      globalThis.VideoEncoder = mockSelf.VideoEncoder;
+      mockSelf.AudioEncoder = originalAudioEncoderCtorForThisTest;
+      globalThis.AudioEncoder = mockSelf.AudioEncoder;
+    });
+  });
+
+});
